requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-filter>=23.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
python-dateutil>=2.8.0
lxml>=4.9.0
mysqlclient>=2.2.0  # Para MySQL (opcional)
coverage>=7.0.0
